requirements.txt
ffmpy3>=0.2.3
python-telegram-bot>=12.7